+++ javaScript/projects/weather-app/src/components/tabs/today/today.config.js
@@ -1,32 +1,46 @@
 export default [
   {
     temp_f: 'Temperature',
+    unit: '°F',
   },
   {
     feelslike_f: 'Feels like',
+    unit: '°F',
   },
   {
     maxtemp_f: 'High',
+    unit: '°F',
   },
   {
     mintemp_f: 'Low',
+    unit: '°F',
   },
   {
     humidity: 'Humidity',
+    unit: '%',
   },
   {
     pressure_in: 'Pressure',
+    unit: 'in',
   },
   {
     vis_miles: 'Visibility',
+    unit: 'mi',
   },
   {
     wind_mph: 'Wind',
+    unit: 'mph',
   },
-  {
-    wind_degree: 'Wind degree',
-  },
+  // {
+  //   wind_degree: 'Wind degree',
+  // },
   {
     wind_dir: 'Wind direction',
+  },
+];
+
+export const details = [
+  {
+    text: null,
   },
 ];