+++ javaScript/projects/weather-app/src/helpers/formatDate.js
@@ -1,0 +1,22 @@
+export default function formatDate(dateString) {
+  // this returns a format [Day, Month Date]
+  // ex: Wednesday, April 14
+  const days = ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'];
+  const months = [
+    'January',
+    'February',
+    'March',
+    'April',
+    'May',
+    'June',
+    'July',
+    'August',
+    'September',
+    'October',
+    'November',
+    'December',
+  ];
+
+  const date = new Date(dateString);
+  return `${days[date.getDay()]}, ${months[date.getMonth()]} ${date.getDate()}`;
+}